Remove the int32 from the indexes of the arrays for insert sort

+++ sort/insertion_sort.go
@@ -1,13 +1,13 @@
 package sort
 
-func swap(nums []int32, i, j int32) {
+func swap(nums []int32, i, j int) {
 	tmp := nums[i]
 	nums[i] = nums[j]
 	nums[j] = tmp
 }
 
 func InsertionSort(nums []int32) {
-	for pos := int32(1); pos < int32(len(nums)); pos++ {
+	for pos := 1; pos < len(nums); pos++ {
 		item := nums[pos]
 
 		for j := pos - 1; j >= 0; j-- {
